Switch Gst source from libcurl to libsoap
curlhttpsrc is in the "bad" plugins list and for a reason! It does have
a lot of problems when it comes to seeking. For example, it sometimes
randomly sends EOS message.

soaphttpsrc on the other hand seems to be solid and I haven't found any
problems with it so far. Conveniently they both have the same API.