Add explicit buffer in the GST pipeline
For some reason this fixes an I/O timeout in libsoup when I am
connecting or disconnecting from a VPN. I am not exactly sure why
this is the case but as long as it works I am happy.

At the same time this commit does not follow the buffering docs in that
I don't pause the playback waiting for the high watermark. But
everything appears to be working as expected so I will not bother yet.